requirements.txt
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
colorama==0.3.9
decorator==4.2.1
Django==1.11
django-allauth==0.34.0
enum34==1.1.6
geopy==1.11.0
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
lxml==4.1.1
oauthlib==2.0.6
pathlib2==2.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
win-unicode-console==0.5
